Significant work: Added two bits to the register code, to keep track of whether use of the register requires a REX prefix, or prohibits it, or can be used with or without a REX prefix.
Register size (width) was not getting set on creation.

Disabled lazy initialization of AJMem size -- the width of the operation can only be known through either the other operand or explicitly being set through #ptr32 and the like.

Added 8-bit registers SPL, BPL, SIL, DIL.

Regularized #isX86, which was incorrectly answering false for MMX registers, and may have given incorrect answers for some XMM registers.

A REX prefix of 0x40 can now be emitted, when it is required (only for registers SPL and so on).

#testCall was expecting REX.W to be set in a number of cases where it's not needed. It's no longer being generated in those cases, probably because of no longer lazy-initializing the size of an AJMem.