Limited victory. My new tests are now passing, but I've made a couple of other tests fail in ways for which the fix is not obvious.
Fixed a number of mistakes in #testByteRegs4through7.

Added accessors for BPL, DIL, SIL, SPL.

Added operand conflict checking at instruction creation time, rather than waiting until code generation. Not all conflicts can be checked early, and not all that can be checked early are checked early at this point.

Fixed a couple of bugs in the emitting of the test instruction -- it wouldn't accept a nil size for the memory operand, and emitted opcode 85 even if it was a byte transfer due to checking the wrong operand for whether it was a byte register.

Not all operands understood #prohibitsRex.